% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/build.cpp
\name{build}
\alias{build}
\title{Return quasardb API build information.}
\usage{
build()
}
\value{
API build information
}
\description{
Return build information of the underlying quasardb C API.

Build information may be useful when a problem or a bug has
been encountered to precisely describe the version of the API used.
}
\examples{
build()
}
