% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp.R
\docType{data}
\name{column_type}
\alias{column_type}
\title{Type of a timeseries column.

Possible types of a timeseries column.}
\format{An object of class \code{list} of length 4.}
\usage{
column_type
}
\description{
Currently, a timeseries can have columns of the following types:
  - double    - column of floating point values.
  - blob      - column of binary data.
  - integer   - column of signed 64-bit integer values.
  - timestamp - column of nanosecond-precision timestamps.
}
\keyword{datasets}
