% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/connect.cpp
\name{connect}
\alias{connect}
\title{Connect to a quasardb cluster.}
\usage{
connect(uri = "qdb://127.0.0.1:2836")
}
\arguments{
\item{uri}{Cluster URI, e.g.: \code{"qdb://127.0.0.1:2836"},
\code{"qdb://10.0.1.1:2836,10.0.1.2:2837,my.database.com:2838"}.}
}
\value{
Handle to the cluster
}
\description{
Connect to a quasardb cluster described with the given URI and returns a
handle to it.
The URI must be in a form \code{"qdb://<ip or hostname>:<port>[, ...]"}.
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
}
