% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/Detach_tags.cpp
\name{detach_tags}
\alias{detach_tags}
\title{Detach tags from an entry.}
\usage{
detach_tags(handle, entry, tags)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}

\item{entry}{An alias (key) of an entry.}

\item{tags}{A vector of tag names to attach.}
}
\description{
Detach one or many tags from an existing entry.

When detaching tags, no error will be raised if the entry has been untagged
from at least one tag (i.e. when this entry had no tags amongst the given
ones).
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
detach_tags(handle, entry = "key", tags = "tag")

detach_tags(handle, entry = "key", tags = c("tag1", "tag2"))
}
\seealso{
\url{https://doc.quasardb.net/master/concepts/tags.html}

\code{\link{connect}}
}
