% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/remove.cpp
\name{entry_remove}
\alias{entry_remove}
\title{Remove an entry.}
\usage{
entry_remove(handle, name)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}

\item{name}{Timeseries name.}
}
\description{
Remove an existing entry from the database.
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
entry_remove(handle, "timeseries")
}
\seealso{
\code{\link{connect}}
}
