% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/error.cpp
\name{error}
\alias{error}
\title{Return error description.}
\usage{
error(code)
}
\arguments{
\item{code}{Error code.}
}
\value{
Error description string.
}
\description{
Return error description string for a given error code.
}
\examples{
error(code = 0)
}
