% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/error.cpp
\name{get_last_error}
\alias{get_last_error}
\title{Return last error.}
\usage{
get_last_error(handle)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}
}
\value{
Last error description.
}
\description{
Return error description of the last error occurred when calling the API.
}
\examples{
get_last_error()
}
