% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/get_tagged.cpp
\name{get_tagged}
\alias{get_tagged}
\title{Get tagged entries.}
\usage{
get_tagged(handle, tag)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}

\item{tag}{An alias (key) of a tag.}
}
\value{
List of entries.
}
\description{
Return keys of all the entries tagged by the given tag.
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
get_tagged(handle, tag = "tag")
}
\seealso{
\code{\link{connect}}
}
