% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/get_tags.cpp
\name{get_tags}
\alias{get_tags}
\title{Get tags.}
\usage{
get_tags(handle, name)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}

\item{name}{An alias (key) of an entry.}
}
\value{
List of tags.
}
\description{
Return all tags of an entry.
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
get_tags(handle, name = "key")
}
\seealso{
\code{\link{connect}}
}
