% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/build.cpp
\name{qdb_build}
\alias{qdb_build}
\title{Return quasardb API build information.

Return build information of the underlying quasardb C API.}
\usage{
qdb_build()
}
\value{
API build information
}
\description{
Build information may be useful when a problem or a bug has
been encountered to precisely describe the version of the API used.
}
\examples{
qdb_build()
}
