% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/Detach_tags.cpp
\name{qdb_detach_tags}
\alias{qdb_detach_tags}
\title{Detach tags from an entry.

Detach one or many tags from an existing entry.}
\usage{
qdb_detach_tags(handle, entry, tags)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{entry}{An alias (key) of an entry.}

\item{tags}{A vector of tag names to attach.}
}
\description{
Tagging is a simple but powerful system that can be used to
make huge quantities of data manageable.
See details, examples and more at
https://doc.quasardb.net/master/concepts/tags.html.

When detaching tags, no error will be raised if the entry has been untagged
from at least one tag (i.e. when this entry had no tags amongst the given
ones).
}
\examples{
qdb_detach_tags(handle, entry = "key", tags = "tag")

qdb_detach_tags(handle, entry = "key", tags = c("tag1", "tag2"))
}
