% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/get_tagged.cpp
\name{qdb_get_tagged}
\alias{qdb_get_tagged}
\title{Return all entries tagged by a tag.}
\usage{
qdb_get_tagged(handle, tag)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{tag}{An alias (key) of a tag.}
}
\value{
List of entries.
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
qdb_get_tagged(handle, tag = "tag")
}
