% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/get_tags.cpp
\name{qdb_get_tags}
\alias{qdb_get_tags}
\title{Return all tags of an entry.}
\usage{
qdb_get_tags(handle, name)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{name}{An alias (key) of an entry.}
}
\value{
List of tags.
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
qdb_get_tags(handle, name = "key")
}
