% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/qdb_query_find.cpp
\name{qdb_query_find}
\alias{qdb_query_find}
\title{Execute a query to find matching entries.}
\usage{
qdb_query_find(handle, query)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{query}{A query to execute.}
}
\value{
List of matching entry aliases/keys.
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
qdb_query_find(handle, "query")
}
