% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/remove.cpp
\name{qdb_remove}
\alias{qdb_remove}
\title{Remove an entry.}
\usage{
qdb_remove(handle, name)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{name}{Timeseries name.}
}
\description{
Remove an existing entry from the database.
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
qdb_remove(handle, "timeseries")
}
