% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/show.cpp
\name{qdb_show}
\alias{qdb_show}
\title{Describe a timeseries.}
\usage{
qdb_show(handle, name)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{name}{A name of the timeseries.}
}
\description{
List columns, their names and types, of a timeseries.
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
columns <- qdb_show(handle, name = "ts")
}
