% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qdb_version}
\alias{qdb_version}
\title{Return quasardb API version}
\usage{
qdb_version()
}
\value{
API version
}
\description{
Return version of the underlying quasardb C API.
}
\examples{
qdb_version()
}
