\name{quasardb-package}
\alias{quasardb-package}
\alias{quasardb}
\docType{package}
\title{
  R API for quasardb database
}
\description{
  The 'quasardb' package provides R API for quasardb timeseries database.
  See doc.quasardb.net for documentation.
  TODO: A length of about one to five lines is recommended.
}
\details{
  TODO: This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Marek Kurdej.

Maintainer: Marek Kurdej <marek@quasardb.net>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{databse}
\seealso{
  TODO: Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
