% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_database.R
\name{run_database}
\alias{run_database}
\title{Helper to run quasardb database server (node).}
\usage{
run_database(root, address = "127.0.0.1:2836")
}
\arguments{
\item{root}{Root directory in which the executable should be found.}

\item{address}{Daemon listening endpoint.}
}
\description{
Run quasardb executable (qdbd) searched for in the given root directory.
Used for testing purposes.
}
