% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/ts_create.cpp
\name{ts_create}
\alias{ts_create}
\title{Create a timeseries.}
\usage{
ts_create(handle, name, columns)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}

\item{name}{A name of the to-be-created timeseries.}

\item{columns}{A named vector of `column_type` integers.}
}
\description{
Create a timeseries with given columns.
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
ts_create(handle, name = "ts",
    columns = c("col1" = column_type$blob, "col2" = column_type$double))
}
\seealso{
\code{\link{connect}}
}
