% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/ts_insert.cpp
\name{ts_double_insert}
\alias{ts_double_insert}
\title{Insert floating-point data.}
\usage{
ts_double_insert(handle, name, column)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}

\item{name}{Timeseries name.}

\item{column}{Column name.}
}
\description{
Insert floating-point data into double column of an existing timeseries.
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
ts_double_insert(handle, "timeseries", "column", points)
}
\seealso{
\code{\link{connect}}
}
