% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/show.cpp
\name{ts_show}
\alias{ts_show}
\title{Describe a timeseries.}
\usage{
ts_show(handle, name)
}
\arguments{
\item{handle}{A valid cluster opened using `connect`.}

\item{name}{A name of the timeseries.}
}
\description{
List columns, their names and types, of a timeseries.
}
\examples{
handle <- connect("qdb://127.0.0.1:2836")
columns <- ts_show(handle, name = "ts")
}
\seealso{
\code{\link{connect}}
}
