% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/version.cpp
\name{version}
\alias{version}
\title{Return quasardb API version}
\usage{
version()
}
\value{
API version
}
\description{
Return version of the underlying quasardb C API.

Version information may be useful when a problem or a bug has
been encountered to precisely describe the version of the API used.
}
\examples{
version()
}
