% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/attach_tags.cpp
\name{qdb_attach_tags}
\alias{qdb_attach_tags}
\title{Attach tags to an entry.

Attach one or many tags to an existing entry.}
\usage{
qdb_attach_tags(handle, entry, tags)
}
\examples{
qdb_attach_tags(handle, 'key', 'tag')
}
