% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qdb_build}
\alias{qdb_build}
\title{Return quasardb API build information.}
\usage{
qdb_build()
}
\value{
API build information
}
\description{
Return build information of the underlying quasardb C API.
}
\examples{
qdb_build()
}
