% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/qdb_build.cpp
\name{qdb_build}
\alias{qdb_build}
\title{Return quasardb API build information.

Return build information of the underlying quasardb C API.}
\usage{
qdb_build()
}
\value{
API build information
}
\examples{
qdb_build()
}
