% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/connect.cpp
\name{qdb_connect}
\alias{qdb_connect}
\title{Connect to a quasardb cluster.

Connect to a quasardb cluster described with the given URI and returns a
  handle to it. The URI must be in a form \code{"qdb://<ip or
  hostname>:<port>[, ...]"}.}
\usage{
qdb_connect(uri)
}
\arguments{
\item{uri}{Cluster URI, e.g.: \code{"qdb://127.0.0.1:2836"},
\code{"qdb://10.0.1.1:2836,10.0.1.2:2837,my.database.com:2838"}.}
}
\value{
Handle to the cluster
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
}
