% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/Detach_tags.cpp
\name{qdb_detach_tags}
\alias{qdb_detach_tags}
\title{Detach tags from an entry.

Detach one or many tags from an existing entry.}
\usage{
qdb_detach_tags(handle, entry, tags)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{entry}{An alias (key) of an entry.}

\item{tags}{A vector of tag names to attach.}
}
