% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qdb_error}
\alias{qdb_error}
\title{Return error description.}
\usage{
qdb_error(ierr)
}
\arguments{
\item{code}{Error code.}
}
\value{
Error description string.
}
\description{
Return error description string for a given error code.
}
\examples{
handle <- qdb_error(0)
}
