% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/error.cpp
\name{qdb_error}
\alias{qdb_error}
\title{Return error description.

Return error description string for a given error code.}
\usage{
qdb_error(ierr)
}
\arguments{
\item{code}{Error code.}
}
\value{
Error description string.
}
\examples{
handle <- qdb_error(0)
}
