% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/query.cpp
\name{qdb_query}
\alias{qdb_query}
\title{Execute a query.

Execute a select query using SQL-like syntax.}
\usage{
qdb_query(handle, query)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{query}{A query to execute.}
}
\value{
A data frame with results of the query.
}
\description{
You can find more details about the possible queries,
the query language syntax, as well as examples and more at
https://doc.quasardb.net/master/api/queries.html.
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
qdb_query(handle, "SELECT * FROM timeseries IN RANGE(2017, +1y)")
}
