% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/ts_create.cpp
\name{qdb_ts_create}
\alias{qdb_ts_create}
\title{Create a timeseries.}
\usage{
qdb_ts_create(handle, name, columns)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{name}{A name of the to-be-created timeseries.}

\item{columns}{A named vector of `column_type` integers.}
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
qdb_ts_create(handle, name = "ts",
    columns = c("col1" = column_type$blob, "col2" = column_type$double))
}
