% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/ts_insert.cpp
\name{qdb_ts_insert.double}
\alias{qdb_ts_insert.double}
\title{Insert floating-point data.}
\usage{
qdb_ts_insert.double(handle, name, column)
}
\arguments{
\item{handle}{A valid cluster opened using qdb_connect.}

\item{name}{Timeseries name.}

\item{column}{Column name.}
}
\description{
Insert floating-point data into double column of an existing timeseries.
}
\examples{
handle <- qdb_connect("qdb://127.0.0.1:2836")
qdb_ts_insert.double(handle, "timeseries", "column", points)
}
